- 更新`Department`/`Permission`/`Variable` - 优化权限     - 删除权限时, 会移除角色 与 此权限 及 子权限的关联 - 去除 修改单位/修改权限 的冗余HTML代码 - 更新所有的校验情况, 进行逻辑整理和代码优化

+++ src/main/java/in/hocg/web/modules/web/admin/system/VariableController.java
@@ -89,7 +89,11 @@ public class VariableController extends BaseController {
     @PostMapping("/delete")
     @ResponseBody
     @PreAuthorize("hasRole('ADMIN')")
-    public Results delete(@Validated IdsFilter filter) {
+    public Results delete(@Validated IdsFilter filter,
+                          BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return Results.check(bindingResult);
+        }
         variableService.delete(filter.getId());
         return Results.success().setMessage("删除成功");
     }
